refactor ffihub
