+++ engine_store_ffi/src/lib.rs
@@ -7,6 +7,7 @@ pub mod interfaces;
 pub mod basic_ffi_impls;
 pub mod domain_impls;
 pub mod encryption_impls;
+pub mod ffihub_impl;
 mod lock_cf_reader;
 pub mod observer;
 pub mod ps_engine;
@@ -28,6 +29,7 @@ pub use domain_impls::*;
 use encryption::DataKeyManager;
 pub use encryption_impls::*;
 use engine_traits::{Peekable, CF_LOCK};
+pub use ffihub_impl::TiFlashFFIHub;
 use kvproto::{kvrpcpb, metapb, raft_cmdpb};
 use lazy_static::lazy_static;
 use protobuf::Message;
